backend/validations.py
from extensions import ma
from models import User, Product

class UserSchema(ma.SQLAlchemyAutoSchema):
    class Meta:
        model = User
        load_instance = True
        # Exclude sensitive info (e.g., password)
        exclude = ("password",)

class ProductSchema(ma.SQLAlchemyAutoSchema):
    class Meta:
        model = Product
        load_instance = True